#17 removendo OutcomeCuiPaleForm (desnecessário)
