--- src/views/Outcome/components/OutcomeCuiPaleForm/index.js
@@ -1,1 +1,0 @@
-export { default } from './OutcomeCuiPaleForm'
